codegen: Add fields and variants to context
